+++ src/randomize/mod.rs
@@ -3,6 +3,7 @@ pub mod antechambers;
 pub mod autoscroll;
 pub mod beta_tornado;
 pub mod bowser_castle;
+pub mod credits;
 pub mod enemies;
 pub mod enemy_protections;
 pub mod fire_flower;